Separate ArrayAccess interface from model to trait
